+++ app/src/main/java/com/assingment/android/view/home/HomeFragment.kt
@@ -7,17 +7,22 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.widget.doAfterTextChanged
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import com.assingment.android.R
+import com.assingment.android.view.NavigationEvent
 import com.assingment.android.view.ShareViewModel
+import com.google.android.material.textfield.TextInputLayout
 
 class HomeFragment: Fragment() {
 
     private val layoutId: Int = R.layout.fragment_home
 
-    private val model: ShareViewModel by activityViewModels()
+    private val viewModel: ShareViewModel by activityViewModels()
+
+    private var nameInputLayout: TextInputLayout? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -31,12 +36,32 @@ class HomeFragment: Fragment() {
         super.onViewCreated(view, savedInstanceState)
         val button = view.findViewById<Button>(R.id.ctaButton)
         val editText = view.findViewById<EditText>(R.id.nameInput)
+        nameInputLayout = view.findViewById(R.id.nameInputLayout)
 
         (activity as AppCompatActivity).supportActionBar?.title = getString(R.string.home_title)
 
         button.setOnClickListener {
-            model.name.value = editText.text.toString()
-            findNavController().navigate(R.id.action_homeFragment_to_quizFragment)
+            viewModel.proceedToQuizScreen(editText.text)
         }
+        editText.doAfterTextChanged {
+            viewModel.onNameInputTextChanged(it)
+        }
+
+        viewModel.apply {
+            nameError.observe(viewLifecycleOwner, ::showNameInputError)
+            navigation.observe(viewLifecycleOwner, ::navigateToDestination)
+        }
+    }
+
+    private fun showNameInputError(message: String?) {
+        nameInputLayout?.apply {
+            error = message
+            isErrorEnabled = !message.isNullOrEmpty()
+        }
+    }
+
+    private fun navigateToDestination(navigation: NavigationEvent?) {
+        navigation ?: return
+        findNavController().navigate(navigation.destination)
     }
 }